refractoring grid hook
change grid hook
