+++ src/components/Field/Field.test.tsx
@@ -1,12 +1,14 @@
 import React from "react";
 import { render } from "@testing-library/react";
+import { GridProvider } from "../../hooks/grid.hook";
+import "@testing-library/jest-dom";
 
 import { Field } from "./Field";
 
 describe("test Field component", () => {
   test("renders Field component", () => {
-    const { container } = render(<Field />);
+    const { getByTestId } = render(<GridProvider><Field /></GridProvider>);
 
-    expect(container).toMatchSnapshot();
+    expect(getByTestId("grid")).toBeInTheDocument();
   });
 });
